+++ app/src/main/java/me/jerryhanks/stepviewapp/MainActivity.kt
@@ -9,7 +9,6 @@ import kotlinx.android.synthetic.main.activity_main.*
 import me.jerryhanks.stepview.IndicatorAdapter
 import me.jerryhanks.stepview.interfaces.TimeLineViewCallback
 import me.jerryhanks.stepview.model.Status
-import java.util.*
 
 class MainActivity : AppCompatActivity() {
 
@@ -17,24 +16,20 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        val timeLines = ArrayList<MyTimeLine>()
-        timeLines.add(MyTimeLine(Status.COMPLETED, getString(R.string.title_1), getString(R.string.content_1)))
-        timeLines.add(MyTimeLine(Status.UN_COMPLETED, getString(R.string.title_2), getString(R.string.content_2)))
-        timeLines.add(MyTimeLine(Status.COMPLETED, getString(R.string.title_3), getString(R.string.content_3)))
-        timeLines.add(MyTimeLine(Status.COMPLETED, getString(R.string.title_4), getString(R.string.content_4)))
-        timeLines.add(MyTimeLine(Status.ATTENTION, getString(R.string.title_5), getString(R.string.content_5)))
-        timeLines.add(MyTimeLine(Status.COMPLETED, getString(R.string.title_6), getString(R.string.content_6)))
+        val timeLines = mutableListOf<MyTimeLine>()
+                .apply {
+                    add(MyTimeLine(Status.COMPLETED, getString(R.string.s_title_1), getString(R.string.s_content_1)))
+                    add(MyTimeLine(Status.COMPLETED, getString(R.string.s_title_2), getString(R.string.s_content_2)))
+                    add(MyTimeLine(Status.COMPLETED, getString(R.string.s_title_3), getString(R.string.s_content_3)))
+                    add(MyTimeLine(Status.COMPLETED, getString(R.string.s_title_4), getString(R.string.s_content_4)))
+                    add(MyTimeLine(Status.COMPLETED, getString(R.string.s_title_5), getString(R.string.s_content_5)))
+                    add(MyTimeLine(Status.COMPLETED, getString(R.string.s_title_6), getString(R.string.s_content_6)))
+                    add(MyTimeLine(Status.COMPLETED, getString(R.string.s_title_7), getString(R.string.s_content_7)))
 
-//
-//        timeLines.add(MyTimeLine(Status.COMPLETED, getString(R.string.s_title_1), getString(R.string.s_content_1)))
-//        timeLines.add(MyTimeLine(Status.UN_COMPLETED, getString(R.string.s_title_2), getString(R.string.s_content_2)))
-//        timeLines.add(MyTimeLine(Status.COMPLETED, getString(R.string.s_title_3), getString(R.string.s_content_3)))
-//        timeLines.add(MyTimeLine(Status.COMPLETED, getString(R.string.s_title_4), getString(R.string.s_content_4)))
-//        timeLines.add(MyTimeLine(Status.ATTENTION, getString(R.string.s_title_5), getString(R.string.s_content_5)))
-//        timeLines.add(MyTimeLine(Status.COMPLETED, getString(R.string.s_title_6), getString(R.string.s_content_6)))
+                }
 
 
-        val adapter = IndicatorAdapter(timeLines, this, object : TimeLineViewCallback<MyTimeLine> {
+        val adapter = IndicatorAdapter(mutableListOf(), this, object : TimeLineViewCallback<MyTimeLine> {
             override fun onBindView(model: MyTimeLine, container: FrameLayout, position: Int): View {
                 val view = layoutInflater
                         .inflate(R.layout.sample_time_line,
@@ -50,8 +45,8 @@ class MainActivity : AppCompatActivity() {
         adapter.swapItems(timeLines)
 
         //set the title
-        caption.text = getString(R.string.timeline_of_world_war_i)
-//        caption.text = getString(R.string.delivery_status)
+//        caption.text = getString(R.string.timeline_of_world_war_i)
+        caption.text = getString(R.string.delivery_status)
 
     }
 }