+++ CHANGELOG.md
@@ -9,4 +9,4 @@ Version 1.0.0 *(2017-11-23)*
 Version 1.0.1 *(2017-12-06)*
 ----------------------------
 
-- imeLineViewCallback callback now returns a view
+- TimeLineViewCallback callback now returns a view